+++ pkg/fuzzer/job_test.go
@@ -25,14 +25,14 @@ func TestDeflakeFail(t *testing.T) {
 	testJob := &triageJob{
 		p:         prog,
 		info:      &flatrpc.CallInfo{},
-		newSignal: signal.FromRaw([]uint32{0, 1, 2, 3, 4}, 0),
+		newSignal: signal.FromRaw([]uint64{0, 1, 2, 3, 4}, 0),
 	}
 
 	run := 0
 	ret, stop := testJob.deflake(func(_ *queue.Request, _ ProgTypes) *queue.Result {
 		run++
 		// For first, we return 0 and 1. For second, 1 and 2. And so on.
-		return fakeResult(0, []uint32{uint32(run), uint32(run + 1)}, []uint32{10, 20})
+		return fakeResult(0, []uint64{uint64(run), uint64(run + 1)}, []uint64{10, 20})
 	}, nil, false)
 	assert.False(t, stop)
 	assert.Equal(t, 5, run)
@@ -51,21 +51,21 @@ func TestDeflakeSuccess(t *testing.T) {
 	testJob := &triageJob{
 		p:         prog,
 		info:      &flatrpc.CallInfo{},
-		newSignal: signal.FromRaw([]uint32{0, 1, 2}, 0),
+		newSignal: signal.FromRaw([]uint64{0, 1, 2}, 0),
 	}
 	run := 0
 	ret, stop := testJob.deflake(func(_ *queue.Request, _ ProgTypes) *queue.Result {
 		run++
 		switch run {
 		case 1:
-			return fakeResult(0, []uint32{0, 2, 4, 6, 8}, []uint32{10, 20})
+			return fakeResult(0, []uint64{0, 2, 4, 6, 8}, []uint64{10, 20})
 		case 2:
 			// This one should be ignored -- it has a different errno.
-			return fakeResult(1, []uint32{0, 1, 2}, []uint32{100})
+			return fakeResult(1, []uint64{0, 1, 2}, []uint64{100})
 		case 3:
-			return fakeResult(0, []uint32{0, 2, 4, 6, 8}, []uint32{20, 30})
+			return fakeResult(0, []uint64{0, 2, 4, 6, 8}, []uint64{20, 30})
 		case 4:
-			return fakeResult(0, []uint32{0, 2, 6}, []uint32{30, 40})
+			return fakeResult(0, []uint64{0, 2, 6}, []uint64{30, 40})
 		}
 		// We expect it to have finished earlier.
 		t.Fatal("only 4 runs were expected")
@@ -73,14 +73,14 @@ func TestDeflakeSuccess(t *testing.T) {
 	}, nil, false)
 	assert.False(t, stop)
 	// Cover is a union of all coverages.
-	assert.ElementsMatch(t, []uint32{10, 20, 30, 40}, ret.cover.Serialize())
+	assert.ElementsMatch(t, []uint64{10, 20, 30, 40}, ret.cover.Serialize())
 	// 0, 2, 6 were in three resuls.
-	assert.ElementsMatch(t, []uint32{0, 2, 6}, ret.stableSignal.ToRaw())
+	assert.ElementsMatch(t, []uint64{0, 2, 6}, ret.stableSignal.ToRaw())
 	// 0, 2 were also in newSignal.
-	assert.ElementsMatch(t, []uint32{0, 2}, ret.newStableSignal.ToRaw())
+	assert.ElementsMatch(t, []uint64{0, 2}, ret.newStableSignal.ToRaw())
 }
 
-func fakeResult(errno int32, signal, cover []uint32) *queue.Result {
+func fakeResult(errno int32, signal, cover []uint64) *queue.Result {
 	return &queue.Result{
 		Info: &flatrpc.ProgInfo{
 			Calls: []*flatrpc.CallInfo{
